♻️ map関数で同じ内容のコンポーネントをまとめる、ファイル名の修正
